Compiladores/Proyecto #1/requirements.txt
antlr4-python3-runtime==4.13.1
pytest==8.3.2
